fix substra tests
